Move require statement to limit dependency
Putting the require statement on the core audiofhile.rb in the
spec_helper was a reflex to solve a failing test with the side-effect of
dependencies being taken for granted. Keeping the require statement
close to the dependency keeps us aware of the dependencies we are adding
a the class.